MOFI/Backend/Producer/backend/requirements.txt
fastapi[standard]
pydantic
uvicorn[standard]
motor
python-dotenv
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1 
pyjwt
cloudinary
python-multipart